// frontend/src/app/core/api/model/create-user-request-dto.ts
/**
 * Service Search App REST API documentation
 * This file holds the REST API documentation of the Hollo Event Notifier project.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateUserRequestDto { 
    username: string;
    email: string;
    role: CreateUserRequestDtoRoleEnum;
    id?: string;
    password: string;
}
export enum CreateUserRequestDtoRoleEnum {
    SystemAdmin = 'SystemAdmin',
    EventAdmin = 'EventAdmin'
};