// frontend/src/app/core/api/model/user-dto.ts
/**
 * Service Search App REST API documentation
 * This file holds the REST API documentation of the Hollo Event Notifier project.
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * This DTo represents a user in the application
 */
export interface UserDto {
    username: string;
    email: string;
    role: UserDtoRoleEnum;
    approved: boolean;
    id?: string;
}
export enum UserDtoRoleEnum {
    SystemAdmin = 'SystemAdmin',
    EventAdmin = 'EventAdmin'
};